move CommentList.tsx to components/CommentList/index.tsx
